firewall_uygulama/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
django-admin==2.0.1
django-excel-base==1.0.5
django-excel-response2==3.0.3
django-six==1.0.4
idna==3.4
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.1
python-http-client==3.3.7
pytz==2022.7.1
requests==2.28.2
screen==1.0.1
sendgrid==6.9.7
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
twilio==7.16.2
urllib3==1.26.14
xlwt==1.3.0
xmltodict==0.13.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework-jsonapi==6.0.0
inflection==0.5.1
tzdata==2022.6
